+++ tsMuxer/movDemuxer.cpp
@@ -136,7 +136,7 @@ struct MOVStreamContext : Track
 
     int ctts_index;
     int ctts_sample;
-    unsigned sample_size;
+    int sample_size;
     unsigned sample_count;
     unsigned keyframe_count;
     unsigned time_scale;
@@ -149,7 +149,7 @@ struct MOVStreamContext : Track
     int width;                  ///< tkhd width
     int height;                 ///< tkhd height
     unsigned bits_per_coded_sample;
-    int channels;
+    unsigned channels;
     int packet_size;
     int sample_rate;
     vector<uint32_t> keyframes;
@@ -166,11 +166,12 @@ class MovParsedAudioTrackData final : public ParsedTrackPrivData
     {
         isAAC = false;
     }
+
     void setPrivData(uint8_t* buff, const int size) override
     {
         m_buff = buff;
         m_size = size;
-        m_aacRaw.m_channels = m_sc->channels;
+        m_aacRaw.m_channels = static_cast<uint8_t>(m_sc->channels);
         m_aacRaw.m_sample_rate = m_sc->sample_rate;
         m_aacRaw.m_id = 1;  // MPEG2
         m_aacRaw.m_profile = 0;
@@ -179,6 +180,7 @@ class MovParsedAudioTrackData final : public ParsedTrackPrivData
         m_aacRaw.m_layer = 0;
         m_aacRaw.m_rdb = 0;
     }
+
     void extractData(AVPacket* pkt, uint8_t* buff, const int size) override
     {
         uint8_t* dst = pkt->data;
@@ -190,7 +192,7 @@ class MovParsedAudioTrackData final : public ParsedTrackPrivData
                 frameSize = m_sc->m_index[m_sc->m_indexCur++];
             if (isAAC)
             {
-                m_aacRaw.m_channels = m_sc->channels;
+                m_aacRaw.m_channels = static_cast<uint8_t>(m_sc->channels);
                 m_aacRaw.buildADTSHeader(dst, frameSize + AAC_HEADER_LEN);
                 memcpy(dst + AAC_HEADER_LEN, buff, frameSize);
                 dst += frameSize + AAC_HEADER_LEN;
@@ -204,9 +206,9 @@ class MovParsedAudioTrackData final : public ParsedTrackPrivData
         }
     }
 
-    unsigned newBufferSize(uint8_t* buff, const unsigned size) override
+    int newBufferSize(uint8_t* buff, const int size) override
     {
-        unsigned left = size;
+        int left = size;
         unsigned i = 0;
         for (; left > 4; ++i)
         {
@@ -331,7 +333,7 @@ class MovParsedH264TrackData : public ParsedTrackPrivData
         }
     }
 
-    unsigned newBufferSize(uint8_t* buff, const unsigned size) override
+    int newBufferSize(uint8_t* buff, const int size) override
     {
         const uint8_t* end = buff + size;
         size_t nalCnt = 0;
@@ -351,7 +353,7 @@ class MovParsedH264TrackData : public ParsedTrackPrivData
         size_t spsPpsSize = 0;
         for (auto& i : spsPpsList) spsPpsSize += i.size() + 4;
 
-        return static_cast<int>(size + spsPpsSize + nalCnt * (4ll - nal_length_size));
+        return static_cast<int>(size + spsPpsSize + nalCnt * (4 - nal_length_size));
     }
 
    protected:
@@ -513,7 +515,7 @@ class MovParsedSRTTrackData final : public ParsedTrackPrivData
         m_timeOffset = endTime;
     }
 
-    unsigned newBufferSize(uint8_t* buff, const unsigned size) override
+    int newBufferSize(uint8_t* buff, const int size) override
     {
         const int64_t stored_sttsCnt = sttsCnt;
         const int64_t stored_sttsPos = sttsPos;
@@ -807,7 +809,7 @@ int MovDemuxer::simpleDemuxBlock(DemuxedData& demuxedData, const PIDSet& accepte
             m_firstDemux = true;
             m_mdat_pos = 0;
         }
-        const auto chunkSize = static_cast<unsigned>(found_moof ? m_mdat_data[m_curChunk].second : next - offset);
+        const auto chunkSize = static_cast<int>(found_moof ? m_mdat_data[m_curChunk].second : next - offset);
         const int trackId = static_cast<int>(chunks[m_curChunk].second);
         auto filterItr = m_pidFilters.find(trackId + 1);
         if (filterItr == m_pidFilters.end() && acceptedPIDs.find(trackId + 1) == acceptedPIDs.end())
@@ -822,7 +824,7 @@ int MovDemuxer::simpleDemuxBlock(DemuxedData& demuxedData, const PIDSet& accepte
             const size_t oldSize = vect.size();
             if (st->parsed_priv_data)
             {
-                if (chunkSize > m_tmpChunkBuffer.size())
+                if (static_cast<size_t>(chunkSize) > m_tmpChunkBuffer.size())
                     m_tmpChunkBuffer.resize(chunkSize);
                 const unsigned readed = get_buffer(m_tmpChunkBuffer.data(), chunkSize);
                 if (readed == 0)
@@ -857,7 +859,7 @@ int MovDemuxer::simpleDemuxBlock(DemuxedData& demuxedData, const PIDSet& accepte
                 if (filterItr != m_pidFilters.end())
                 {
                     m_filterBuffer.resize(chunkSize);
-                    const unsigned readed = get_buffer(m_filterBuffer.data(), chunkSize);
+                    const int readed = static_cast<int>(get_buffer(m_filterBuffer.data(), chunkSize));
                     if (readed < chunkSize)
                         m_filterBuffer.grow(readed - chunkSize);
                     if (readed == 0)
@@ -870,7 +872,7 @@ int MovDemuxer::simpleDemuxBlock(DemuxedData& demuxedData, const PIDSet& accepte
                 else
                 {
                     vect.grow(chunkSize);
-                    const unsigned readed = get_buffer(vect.data() + oldSize, chunkSize);
+                    const int readed = static_cast<int>(get_buffer(vect.data() + oldSize, chunkSize));
                     if (readed < chunkSize)
                     {
                         vect.grow(readed - chunkSize);
@@ -1474,7 +1476,7 @@ int MovDemuxer::mov_read_stsd(MOVAtom atom)
                 {
                     get_be32();                                                  // sizeof struct only
                     st->sample_rate = static_cast<int>(av_int2dbl(get_be64()));  // float 64
-                    st->channels = static_cast<int>(get_be32());
+                    st->channels = get_be32();
                     get_be32();                              // always 0x7F000000
                     st->bits_per_coded_sample = get_be32();  // bits per channel if sound is uncompressed
                     get_be32();                              // lcpm format specific flag
