+++ tsMuxer/singleFileMuxer.h
@@ -12,7 +12,7 @@ class SingleFileMuxer final : public AbstractMuxer
     SingleFileMuxer(MuxerManager* owner);
     ~SingleFileMuxer() override;
     bool muxPacket(AVPacket& avPacket) override;
-    void intAddStream(const std::string& streamName, const std::string& codecName, uint16_t streamIndex,
+    void intAddStream(const std::string& streamName, const std::string& codecName, int streamIndex,
                       const std::map<std::string, std::string>& params, AbstractStreamReader* codecReader) override;
     bool doFlush() override;
     bool close() override;
